bugfix: 1등 2번 당첨시 오버플로우 발생 문제 픽스

+++ src/main/java/lotto/domain/LottoGame.java
@@ -22,7 +22,7 @@ public class LottoGame {
         this.money = money;
     }
 
-    public static LottoGame of(final int money) {
+    public static LottoGame of(final long money) {
 
         return new LottoGame(Money.of(money));
     }
